chore: safeCall 분리
